+++ src/ScottPlot5/ScottPlot5 Tests/RenderTests/FontTests.cs
@@ -1,0 +1,15 @@
+﻿namespace ScottPlotTests.RenderTests;
+
+internal class FontTests
+{
+    [Test]
+    public void Test_Font_SpecialCharacters()
+    {
+        FontService.SetDefaultFont('频');
+
+        ScottPlot.Plot plt = new();
+        plt.TopAxis.Label.Text = "频率信号削减123";
+        Console.WriteLine(plt.TopAxis.Label.Font.Name);
+        plt.SaveTestImage();
+    }
+}