+++ src/ScottPlot5/ScottPlot5/FontService.cs
@@ -1,26 +1,27 @@
 ﻿namespace ScottPlot;
 
 /// <summary>
-/// This class provides cross-platform tools for working with fonts
+/// Cross-platform tools for working with fonts
 /// </summary>
 public static class FontService
 {
-    public static SKTypeface SansTypeface { get; private set; } = GetDefaultSansFont();
-    public static SKTypeface MonospaceTypeface { get; private set; } = GetDefaultMonospaceFont();
-    public static SKTypeface SerifTypeface { get; private set; } = GetDefaultSerifFont();
+    /// <summary>
+    /// This default font is used for almost all text rendering
+    /// </summary>
+    public static string DefaultFontName { get; private set; } = GetDefaultSansFont().FamilyName;
 
-    public static string SansFontName => SansTypeface.FamilyName;
-    public static string MonospaceFontName => MonospaceTypeface.FamilyName;
-    public static string SerifFontName => SerifTypeface.FamilyName;
+    public static string SansFontName { get; private set; } = GetDefaultSansFont().FamilyName;
+    public static string SerifFontName { get; private set; } = GetDefaultSerifFont().FamilyName;
+    public static string MonospaceFontName { get; private set; } = GetDefaultMonospaceFont().FamilyName;
 
-    public static string[] GetInstalledFonts()
+    private static HashSet<string> GetInstalledFonts()
     {
-        return SKFontManager.Default.FontFamilies.ToArray();
+        return new(SKFontManager.Default.FontFamilies, StringComparer.InvariantCultureIgnoreCase);
     }
 
-    private static SKTypeface GetDefaultSansFont()
+    public static SKTypeface GetDefaultSansFont()
     {
-        HashSet<string> installedFonts = new(SKFontManager.Default.FontFamilies);
+        var installedFonts = GetInstalledFonts();
 
         string[] preferredFonts = { "Open Sans", "Segoe UI", "Lato", "DejaVu Sans", "Helvetica" };
         foreach (string preferredFont in preferredFonts)
@@ -34,9 +35,9 @@ public static class FontService
         return SKTypeface.Default;
     }
 
-    private static SKTypeface GetDefaultMonospaceFont()
+    public static SKTypeface GetDefaultMonospaceFont()
     {
-        HashSet<string> installedFonts = new(SKFontManager.Default.FontFamilies);
+        var installedFonts = GetInstalledFonts();
 
         string[] preferredFonts = { "Roboto Mono", "Consolas", "DejaVu Sans Mono", "Courier" };
         foreach (string preferredFont in preferredFonts)
@@ -50,9 +51,9 @@ public static class FontService
         return SKTypeface.Default;
     }
 
-    private static SKTypeface GetDefaultSerifFont()
+    public static SKTypeface GetDefaultSerifFont()
     {
-        HashSet<string> installedFonts = new(SKFontManager.Default.FontFamilies);
+        var installedFonts = GetInstalledFonts();
 
         string[] preferredFonts = { "Times New Roman", "DejaVu Serif", "Times" };
         foreach (string preferredFont in preferredFonts)
@@ -64,5 +65,45 @@ public static class FontService
         }
 
         return SKTypeface.Default;
+    }
+
+    /// <summary>
+    /// Return the name of the font which will best display the given character.
+    /// This method helps identify the best fonts for displaying Chinese, Japanese, and Korean (CJK) characters.
+    /// </summary>
+    public static string DetectFont(char c)
+    {
+        // https://github.com/ScottPlot/ScottPlot/issues/2746
+        return SKFontManager.Default.MatchCharacter(c).FamilyName;
+    }
+
+    /// <summary>
+    /// Set the default font to the best one for displaying the given character.
+    /// This method helps display Chinese, Japanese, and Korean (CJK) characters.
+    /// </summary>
+    public static void SetDefaultFont(char c)
+    {
+        // https://github.com/ScottPlot/ScottPlot/issues/2746
+        DefaultFontName = DetectFont(c);
+    }
+
+    /// <summary>
+    /// Set the default font to the best one for displaying the given character.
+    /// This method helps display Chinese, Japanese, and Korean (CJK) characters.
+    /// </summary>
+    public static void SetDefaultFont(string fontName, bool throwIfNotFound = false)
+    {
+        var installedFonts = GetInstalledFonts();
+
+        if (!installedFonts.Contains(fontName))
+        {
+            DefaultFontName = fontName;
+            return;
+        }
+
+        if (throwIfNotFound)
+        {
+            throw new InvalidOperationException($"Font not found: {fontName}");
+        }
     }
 }